requirements.txt
PyQt6>=6.7.0
yt-dlp>=2026.1.1
PyYAML>=6.0.1
plyer>=2.1.0
